Refactor git_tag operation to be under generic git tools instead of coupling it on the dotnet project.
